requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0
yt_dlp>=2023.0.0
